requirements.txt
cffi==1.16.0
clr-loader==0.2.6
decorator==5.1.1
numpy==1.26.4
psutil==5.9.8
py-cpuinfo==9.0.0
pycparser==2.22
pypiwin32==223
pythonnet==3.0.3
pyuac==0.0.3
pywin32==306
tee==0.0.3
WMI==1.5.1
